move script into its own module
